move models and variables list to constants.py
