+++ finch/processes/constants.py
@@ -1,0 +1,41 @@
+MODELS = [
+    "BNU-ESM",
+    "CCSM4",
+    "CESM1-CAM5",
+    "CNRM-CM5",
+    "CSIRO-Mk3-6-0",
+    "CanESM2",
+    "FGOALS-g2",
+    "GFDL-CM3",
+    "GFDL-ESM2G",
+    "GFDL-ESM2M",
+    "HadGEM2-AO",
+    "HadGEM2-ES",
+    "IPSL-CM5A-LR",
+    "IPSL-CM5A-MR",
+    "MIROC-ESM-CHEM",
+    "MIROC-ESM",
+    "MIROC5",
+    "MPI-ESM-LR",
+    "MPI-ESM-MR",
+    "MRI-CGCM3",
+    "NorESM1-M",
+    "NorESM1-ME",
+    "bcc-csm1-1-m",
+    "bcc-csm1-1",
+]
+
+# a list of all posible netcdf arguments in xclim
+xclim_netcdf_variables = {
+    "tasmin",
+    "tasmax",
+    "tas",
+    "pr",
+    "prsn",
+    "tn10",
+    "tn90",
+    "t10",
+    "t90",
+}
+
+bccaq_variables = {"tasmin", "tasmax", "pr"}
